backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.8,<3.0
twilio>=9.3,<10.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
PyMuPDF>=1.24,<2.0
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0
sentence-transformers>=3.0,<4.0
qdrant-client>=1.10,<2.0
openai>=1.0,<2.0
langgraph>=0.3,<1.0
langchain-core>=0.3,<1.0
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
pymysql>=1.1,<2.0
cryptography>=42.0,<44.0
pdfplumber>=0.11,<1.0
pytesseract>=0.3,<1.0
Pillow>=10.0,<12.0
httpx>=0.27,<1.0
